[#169] move Jackson YAML dependencies to client
