+++ app-swing/build.gradle
@@ -23,10 +23,6 @@ dependencies {
 
   implementation 'com.formdev:flatlaf:3.6'
   implementation 'com.formdev:flatlaf-intellij-themes:3.6'
-
-  implementation platform('com.fasterxml.jackson:jackson-bom:2.19.4')
-  implementation 'com.fasterxml.jackson.core:jackson-databind'
-  implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
 }
 
 test {
