+++ core/build.gradle
@@ -15,9 +15,4 @@ dependencies {
   implementation 'org.xerial:sqlite-jdbc:3.50.1.0'
   implementation 'com.github.ben-manes.caffeine:caffeine:3.2.1'
   implementation 'org.apache.fury:fury-core:0.10.3'
-
-  testImplementation platform('com.fasterxml.jackson:jackson-bom:2.19.4')
-  testImplementation 'com.fasterxml.jackson.core:jackson-databind'
-  testImplementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
-
 }
